Factor out common ripple code/styles
