+++ src/components/mx-toggle-button/mx-toggle-button.tsx
@@ -1,4 +1,5 @@
 import { Component, Host, h, Prop } from '@stencil/core';
+import ripple from '../ripple';
 
 @Component({
   tag: 'mx-toggle-button',
@@ -22,44 +23,13 @@ export class MxToggleButton {
       return;
     }
 
-    this.ripple(e);
+    ripple(e, this.spanElem);
 
     // Allow unchecking of radio button
     if (this.inputElem.checked) {
       this.inputElem.checked = false;
       e.preventDefault();
     }
-  }
-
-  ripple(e: MouseEvent) {
-    let existingRipple = this.spanElem.querySelector('.ripple');
-    if (existingRipple) existingRipple.remove();
-
-    // Create span element
-    let ripple = document.createElement('span');
-
-    // Add ripple class to span
-    ripple.classList.add('ripple');
-
-    // Add span to the button
-    this.spanElem.prepend(ripple);
-
-    // Set the size of the span element
-    const diameter = Math.max(this.spanElem.clientWidth, this.spanElem.clientHeight);
-    ripple.style.width = ripple.style.height = diameter + 'px';
-
-    // Position the span element
-    const elemOffset = this.spanElem.getBoundingClientRect();
-    // Center over click coords OR over top left corner if activated by keypress
-    const left = Math.max(e.clientX - elemOffset.left, 0);
-    const top = Math.max(e.clientY - elemOffset.top, 0);
-    ripple.style.left = left - diameter / 2 + 'px';
-    ripple.style.top = top - diameter / 2 + 'px';
-
-    // Remove span after 0.3s
-    setTimeout(() => {
-      ripple.remove();
-    }, 300);
   }
 
   render() {
