+++ src/tailwind/mx-button/index.scss
@@ -6,22 +6,6 @@
     cursor: pointer;
     width: 100%;
 
-    span.ripple {
-      position: absolute;
-      border-radius: 50%;
-      background-color: var(--ripple-color);
-      animation: ripple 1s;
-      opacity: 1;
-      transform: scale(0);
-    }
-
-    @keyframes ripple {
-      to {
-        opacity: 0;
-        transform: scale(10);
-      }
-    }
-
     &.contained {
       background-color: var(--mds-primary-bg-dark);
       color: var(--mds-primary-text-light);
@@ -30,6 +14,10 @@
       padding: 10px 16px;
       font-weight: 600;
       --ripple-color: rgba(255, 255, 255, 0.32);
+
+      &:focus {
+        background: #3874ba;
+      }
 
       &:hover {
         background: var(--mds-primary-bg-dark);
@@ -52,13 +40,13 @@
       font-weight: 600;
       --ripple-color: rgba(51, 51, 51, 0.12);
 
+      &:focus:not([aria-disabled]) {
+        background: rgba(51, 51, 51, 0.12);
+      }
+
       &:hover:not([aria-disabled]) {
         background: #f0f0f0;
         text-decoration: none !important;
-      }
-
-      &:focus:not([aria-disabled]) {
-        background: rgba(51, 51, 51, 0.12);
       }
 
       &:active:not([aria-disabled]) {
@@ -91,12 +79,12 @@
         color: #c1c1c1;
       }
 
-      &:hover:not([aria-disabled]) {
-        background: #f0f0f0;
-      }
-
       &:focus:not([aria-disabled]) {
         background: #dcdcdc;
+      }
+
+      &:hover:not([aria-disabled]) {
+        background: #f0f0f0;
       }
 
       &:active:not([aria-disabled]) {
@@ -129,17 +117,17 @@
         color: #c1c1c1;
       }
 
-      &:hover:not([aria-disabled]) {
-        background: #ebf2f9;
-        .separator {
-          background: rgba(4, 87, 175, 0.08);
-        }
-      }
-
       &:focus:not([aria-disabled]) {
         background: #c3d7ec;
         .separator {
           background: rgba(4, 87, 175, 0.24);
+        }
+      }
+
+      &:hover:not([aria-disabled]) {
+        background: #ebf2f9;
+        .separator {
+          background: rgba(4, 87, 175, 0.08);
         }
       }
 
@@ -158,6 +146,13 @@
         color: #c1c1c1;
       }
 
+      &:focus:not([aria-disabled]) {
+        background: #d7d7d7;
+        .chevron-wrapper {
+          background: #dcdcdc;
+        }
+      }
+
       &:hover:not([aria-disabled]) {
         background: #eeeeee;
         .chevron-wrapper {
@@ -165,7 +160,6 @@
         }
       }
 
-      &:focus:not([aria-disabled]),
       &:active:not([aria-disabled]) {
         background: #d7d7d7;
         .chevron-wrapper {
