+++ src/tailwind/mx-toggle-button/index.scss
@@ -47,20 +47,4 @@
   &:first-of-type:last-of-type .btn-toggle {
     border-radius: 4px;
   }
-
-  span.ripple {
-    position: absolute;
-    border-radius: 50%;
-    background-color: var(--ripple-color);
-    animation: ripple 1s;
-    opacity: 1;
-    transform: scale(0);
-  }
-
-  @keyframes ripple {
-    to {
-      opacity: 0;
-      transform: scale(10);
-    }
-  }
 }
