+++ src/tailwind/ripple.scss
@@ -1,0 +1,19 @@
+:root {
+  --ripple-color: rgba(255, 255, 255, 0.32);
+}
+
+span.ripple {
+  position: absolute;
+  border-radius: 50%;
+  background-color: var(--ripple-color);
+  animation: ripple 1s;
+  opacity: 1;
+  transform: scale(0);
+}
+
+@keyframes ripple {
+  to {
+    opacity: 0;
+    transform: scale(10);
+  }
+}
